update: dependencies && ignore *.db file
