+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "python.analysis.typeCheckingMode": "off",
+  "githubPullRequests.ignoredPullRequestBranches": [
+    "main"
+  ]
+}
