+++ cogs/yabagame.py
@@ -1,8 +1,6 @@
-from discord import message
-from discord.ext import commands 
-from datetime import date,datetime
-from discord.message import flatten_handlers
-import emoji, discord, asyncio, os, random
+import discord
+from discord.ext import commands
+
 
 class smallgame(commands.Cog):
     def __init__(self, app):
@@ -10,11 +8,13 @@ class smallgame(commands.Cog):
 
     @commands.command(name="야바위")
     async def test(self, ctx):
-        embed = discord.Embed(title="야바위게임", description="⚾　　🏀　　⚽\n\n　　고르세요!", color=0x62c1cc)
+        embed = discord.Embed(
+            title="야바위게임", description="⚾　 🏀　 ⚽\n\n  고르세요!", color=0x62c1cc)
         msg = await ctx.channel.send(embed=embed)
         mmm = await msg.add_reaction("⚾")
         await msg.add_reaction("🏀")
         await msg.add_reaction("⚽")
+
 
 def setup(app):
     app.add_cog(smallgame(app))
