+++ user_data.py
@@ -1,4 +1,3 @@
-
 import sqlite3
 
 conn = sqlite3.connect("user.db", isolation_level=None)
@@ -11,12 +10,12 @@ c.execute("CREATE TABLE IF NOT EXISTS UserData \
 a = []
 money = []
 
+
 class User_Data:
 
     def sign(name, id, money):
         c.execute(f"INSERT INTO UserData \
         VALUES('{name}', '{id}', {money})")
-        
 
     def check(id):
         c.execute("SELECT id FROM UserData")
@@ -31,19 +30,13 @@ class User_Data:
         c.execute("SELECT * FROM UserData WHERE id ='{}'".format(id))
         for row in c.fetchall():
             a.append(row[1])
-        
+
     def change_user(name, id, money):
-        c.execute("DELETE FROM UserData WHERE id = :ID",{"ID":id})
-        User_Data.sign(name, id , money)
+        c.execute("DELETE FROM UserData WHERE id = :ID", {"ID": id})
+        User_Data.sign(name, id, money)
 
     def Selete_money(id):
         c.execute("SELECT * FROM UserData WHERE id ='{}'".format(id))
         for row in c.fetchall():
             money.append(row[2])
             return row[2]
-        
-
-
-
-    
-
